move `test_proto_no_pkg` to proto_reader.rs
