+++ src/proto_reader.rs
@@ -174,4 +174,14 @@ mod test_proto_config {
             None
         }
     }
+    #[tokio::test]
+    async fn test_proto_no_pkg() -> Result<()> {
+        let runtime = crate::runtime::test::init(None);
+        let reader = ProtoReader::init(runtime);
+        let mut proto_no_pkg = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+        proto_no_pkg.push("src/grpc/tests/proto_no_pkg.graphql");
+        let config_module = reader.read(proto_no_pkg.to_str().unwrap()).await;
+        assert!(config_module.is_err());
+        Ok(())
+    }
 }
